Define typed interface for docker
The interface to interact with the 'docker' program now has a type-safe
implementation, using free and typed-process. This allows docker
interaction to be performed with a dedicated monad, meaning

  * Docker commands are represented using functions in the monad rather
    than string literals

  * The input and output of docker functions must be the correct type,
    preventing invalid commands being constructed

  * The necessary plumbing to manage processes is hidden functions which
    call docker commands

To allow actions which use docker to be typed, a DockerT monad
transformer is exposed. This allows docker actions to be composed with
other monadic effects.