Configuraton changes for Postgres
The configuration file format is now changed to use a postgres
configuration type. This includes the a name for the configuration, the
host, the port number and a map of usernames and passwords.

This new style of configuration will eventually phase out the older
"postgres_password" and "db_passwords" fields in the configuration file
allowing configuration for a single database to be stored in the same
key in the configuration file.