Environment variable support in config file
Data in the config file (.dampfcfg.yaml) can now be passed from the
environment. This is done by prefixing the name of the environment
variable with a the dollar sign, e.g.

    postgres:
        host: $PG_HOST
        port: $PG_PORT

        users:
            $PG_USER: $PG_PASS

If an environment variable referenced in the configuration file does not
exist, then an error is thrown which crashes the program. It may be
desirable to attempt recovery using a reasonable default value for
fields with invalid variables.