Pretty printing and updated dump
Pretty printing modules have been defined for the data types in
AppFile.hs and ConfigFile.hs, and the dump command in the program has
been updated to use a pretty printed output. The dumpyaml and dump
commands in the application have been merged into a single dump command
which outputs information about both the application and configuration.