Restructure project
The project has been restructured to use an internal module, which
contains modules which should not be exported by the library. The extra
modules for application and configuration files have been moved to the
internal module, as the toplevel module for each exports all required
functionality.

The cabal file for the dampfs has been reformatted to be more
consistent, and additional compiler warnings have been enabled.