Define dampf monad stack
A monad stack is defined which will eventually be used internally for
all dampf operations. The monad represents an environment with an
immutable application and configuration, and the ability to throw,
catch and mask exceptions in operations. An associated exception type
is also defined which will define exceptions which can be encountered.