YAML utilities module
Identical implementations of ToJSON and FromJSON have been moved to a
YAML utilities module. This means that any lens types which implements
these typeclasses can use the generic definitions which handle the
underscore in identifiers.

The parsing functions for application and configuration have also been
moved into here as a single function, which works on both types.