Redefine application type
The type of application files has been changed from a list of Dampf to
a new type, which is a product of different maps. This offers some
advantages over the previous definition:

  * Filtering by specification type is no longer needed
  * Traversing over particular specifications is easier

Lens instances have been defined for the new type, and an instance for
HasDampfApp has been specified for the DampfContext type.