Support for choosing a profile
It is now possible to have multiple profiles described in the
configuration file, and specify the profile to use when running the
program. This determines which configuration is used when running
actions in DampfT.

Currently there are some limitations in the implementation. Both the
old configuration format and the new one with multiple profiles are
supported, however the key of the root object in the configuration
file must be "Left" or "Right" to determine which format to use.