Handle docker arguments in new module
A new module has been defined to represent arguments for different
docker commands and serialise them into an array of strings for use
with the typed-process library. This allows configuration for more
complex commands (i.e. docker run) to be moved to a separate module,
keeping the related interpreter function smaller and easier to
understand.